+++ website/js/example.js
@@ -4,37 +4,42 @@
   var width   = canvas.width = window.innerWidth;
   var height  = canvas.height = window.innerHeight;
 
-  var mainCircle   = new TYPE6.Circle( width * 0.5, height * 0.5, 200 );
+  var mainCircle   = new Type6.Circle( width * 0.5, height * 0.5, 200 );
   var smallCircles = [];
   var angle        = 0;
   var numObjects   = 24;
   var step         = 0;
-  var minAlpha     = 0.2;
-  var slice        = TYPE6.Trigonometry.twopi / numObjects;
+  var slice        = Type6.Trigonometry.twopi / numObjects;
   var radius       = 20;
+
+  function smallCircle(posX,posY) {
+    this.circle = new Type6.Circle(posX, posY, 20);
+    this.alpha = 0.2;
+    this.minAlpha = 0.2;
+    this.setAlpha = function() {
+      if( this.alpha > this.minAlpha ) {
+        this.alpha = Math.max(this.alpha - 0.01, this.minAlpha);
+      }
+    };
+  }
 
   for(var i = 0; i < numObjects; i += 1) {
     angle = i * slice;
-    smallCircles[i] = new TYPE6.Circle(
-      TYPE6.Trigonometry.cosineEquation( mainCircle.radius, angle, 0, mainCircle.position.x ),
-      TYPE6.Trigonometry.sineEquation( mainCircle.radius, angle, 0, mainCircle.position.y ),
-      20
+    smallCircles[i] = new smallCircle(
+      Type6.Trigonometry.cosineEquation( mainCircle.radius, angle, 0, mainCircle.position.x ),
+      Type6.Trigonometry.sineEquation( mainCircle.radius, angle, 0, mainCircle.position.y )
     );
-    var circle = smallCircles[i];
-    circle.alpha = minAlpha;
-    circle.majAlpha = function(){
-      if( this.alpha > minAlpha )
-        this.alpha = Math.max(this.alpha - 0.01, minAlpha);
-    };
   }
+
 
   function draw(){
     for(var i = 0 ; i < numObjects ; i += 1) {
       var circle = smallCircles[i];
-      if (i === Math.floor(step))
+      if (i === Math.floor(step)){
         circle.alpha = 1;
-      else
-        circle.majAlpha();
+      }else{
+        circle.setAlpha();
+      }
       drawCircle(smallCircles[i]);
     }
 
@@ -47,7 +52,7 @@
   function drawCircle(circle){
     context.fillStyle = "rgba(153, 0, 0, " + circle.alpha + ")";
     context.beginPath();
-    context.arc( circle.position.x, circle.position.y, circle.radius, 0, TYPE6.Trigonometry.twopi, false);
+    context.arc( circle.circle.position.x, circle.circle.position.y, circle.circle.radius, 0, Type6.Trigonometry.twopi, false);
     context.fill();
   }
 
@@ -65,7 +70,7 @@
   }
 
   //create animation frame
-  var animation = new FRAMERAT.Player(render);
+  var animation = new Framerat.Player(render);
 
   function playAnimation(){
     var state = animation.toggle();
@@ -84,7 +89,7 @@
   }
 
   function majTime() {
-    findById('time').innerHTML = formatTime(TYPE6.Utils.round(animation.getTotal(), 2));
+    findById('time').innerHTML = formatTime(Type6.Utils.round(animation.getTotal(), 2));
   }
 
   function majFPS() {