+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Version 0.3.4 (July 15th 2018)
+------------------------------
+ * Library exported as ES6 and IIFE modules instead of UMD.
+ * TYPE6 namespace becomes Type6
+
 Version 0.3.3 (July 4th 2018)
 ------------------------------
  * Documentation automatically generated in /doc folder